requirements.txt
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.12
albumentations>=1.4.4
opencv-python>=4.9.0
scikit-learn>=1.3.2
tensorboard>=2.15.1
pyyaml>=6.0.1
tqdm>=4.66.1
numpy>=1.26.2